add kill, withdraw functions
